==> 2_Data_Preprocessing.ipynb <==
#Data Preprocessing

##Merge Datasets

The first thing we will do is combine the r/India and the r/IndiaSpeaks datasets into one, using the very simple ```pd.concat()``` method

df_india_path = '/content/drive/MyDrive/Wysa/df_india.pkl'
df_ispeaks_path = '/content/drive/MyDrive/Wysa/df_ispeaks.pkl'

import pandas as pd

df_india = pd.read_pickle(df_india_path)
df_ispeaks = pd.read_pickle(df_ispeaks_path)
merged = pd.concat([df_india, df_ispeaks])
merged.head()

merged.info()

##Reduce Dataset

We have a staggering 97 columns of data, ranging from very important textual attributes, to scores and number of comments per post, to trivial internal reddit metadata. Below, I select 17 important columns that we may require in the future for our exploratory data analysis:

| Fields                  	|
|-------------------------	|
| 'author',               	|
|  'created_utc',         	|
|  'full_link',           	|
|  'id',                  	|
|  'is_original_content', 	|
|  'is_self',             	|
|  'is_video',            	|
|  'link_flair_text',     	|
|  'num_comments',        	|
|  'over_18',             	|
|  'score',               	|
|  'selftext',            	|
|  'send_replies',        	|
|  'subreddit',           	|
|  'title',               	|
|  'upvote_ratio',        	|
|  'url'                  	|

df_reduced = merged[["author","created_utc","full_link","id","is_original_content","is_self","is_video","link_flair_text","num_comments","over_18","score","selftext","send_replies","subreddit","title","upvote_ratio","url"]]

df_reduced.head()

df_reduced.info()

##Fill Null Values

The following fields have one or more null values:

is_original_content, is_video, link_flair_text, selftext, send_replies, upvote_ratio

One by one, we will fill these null values depending on the distribution of the non-null values

df_reduced["is_original_content"].value_counts()

Since most are False, we can just do a simple forward fill for wherever there is a null value

df_reduced["is_original_content"] = df_reduced["is_original_content"].fillna(method="ffill")

df_reduced.info()

df_reduced["is_original_content"].value_counts()

df_reduced["is_video"].value_counts()

df_reduced["is_video"] = df_reduced["is_video"].fillna(method="ffill")

df_reduced["is_video"].value_counts()

df_reduced["link_flair_text"].value_counts()

We fill null values for flair text as "NOFLAIR"

df_reduced["link_flair_text"] = df_reduced["link_flair_text"].fillna("NOFLAIR")

df_reduced["selftext"] = df_reduced["selftext"].fillna("NOTEXT")

df_reduced["send_replies"].value_counts()

df_reduced["send_replies"] = df_reduced["send_replies"].fillna(method="ffill")

We fill null entries in upvote_ratio as the mean upvote ratio

df_reduced['upvote_ratio'].describe()

df_reduced["upvote_ratio"] = df_reduced["upvote_ratio"].fillna(0.991904)

df_reduced.info()

Next, we create a new column for text that is a concatenation of the title of a post and the text of the post.

df_reduced["fulltext"] = df_reduced["title"]+' '+df_reduced["selftext"]

df_reduced.head()

df_reduced["fulltext"].iloc[0]

##Date Addition

Next, we create columns for getting human readable date formats

from datetime import datetime
def utc_to_datetime(data):
    data['year'] = data['created_utc'].apply(lambda utc: datetime.fromtimestamp(utc).year)
    data['month'] = data['created_utc'].apply(lambda utc: datetime.fromtimestamp(utc).month)
    data['day'] = data['created_utc'].apply(lambda utc: datetime.fromtimestamp(utc).day)    
    data['hour'] = data['created_utc'].apply(lambda utc: datetime.fromtimestamp(utc).hour)
    data['minute'] = data['created_utc'].apply(lambda utc: datetime.fromtimestamp(utc).minute)
    data['second'] = data['created_utc'].apply(lambda utc: datetime.fromtimestamp(utc).second)

utc_to_datetime(df_reduced)
df_reduced.head()

##Text Preprocessing

Finally, we perform clean up on the fulltext column. We remove links, stopwords, punctuations, extra whitespace, etc. We also create an additional column for lemmatized text using the WordNetLemmatizer.

import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.tag import pos_tag
from nltk.stem import WordNetLemmatizer
import tqdm
import string

def get_wordnet_tag(tag):
    if tag == 'ADJ':
        return 'j'
    elif tag == 'VERB':
        return 'v'
    elif tag == 'NOUN':
        return 'n'
    elif tag == 'ADV':
        return 'r'
    else:
        return 'n'

def lemmatize_text(title):
    stop = stopwords.words('english')
    lemmatizer = WordNetLemmatizer()

    words = list()
    title = word_tokenize(title)
    for word, tag in pos_tag(title):
        tag = get_wordnet_tag(tag)
        word = lemmatizer.lemmatize(word, tag)
        if word not in stop:
            words.append(word)
    
    return ' '.join(words)  

def clean_text(dataset):
    
    tqdm.tqdm.pandas()
    
    dataset['ctext'] = dataset['fulltext'].str.lower()
    dataset['ctext'] = dataset['ctext'].str.replace(r'\[oc\]', ' ')
    pattern_link = r'https?://[^\s]+|www\.[^\s]+|[^\s]+\.com[^\s]*|[^\s]+\.org[^\s]*|[^\s]+\.html[^\s]*'
    dataset['ctext'] = dataset['ctext'].str.replace(pattern_link, ' link ')
    
    pattern_punctuation = r'[' + string.punctuation + '’]'
    dataset['ctext'] = dataset['ctext'].str.replace(pattern_punctuation, '')
    dataset['ctext'] = dataset['ctext'].str.replace(r' [\d]+ |^[\d]+ | [\d]+$', ' ')
    dataset['ctext'] = dataset['ctext'].str.replace(r'[^\w\d\s]+', ' ')
    dataset['lemma_text'] = dataset['ctext'].progress_apply(lambda text: lemmatize_text(text))
    dataset['ctext'] = dataset['ctext'].str.replace(r'\s[\s]+', ' ')
    dataset['lemma_text'] = dataset['lemma_text'].str.replace(r'\s[\s]+', ' ')

nltk.download('stopwords')
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
nltk.download('wordnet')

clean_text(df_reduced)

df_reduced['lemma_text'].iloc[0]

df_reduced.info()

df_reduced.describe()

df_reduced.drop(['url', 'id'], axis=1, inplace=True)

df_reduced.reset_index(drop=True, inplace=True)

df_reduced.to_pickle("/content/drive/MyDrive/Wysa/df_preproc.pkl")
