==> 4_Predictor_Model.ipynb <==
#Predictive Model Creation

In this notebook, we create predictive models for the topic distribution we figured out in the previous notebook. We establish one classical ML baseline as well as a neural baseline using DistilBERT.

However, before proceeding, we must discuss the validity of this approach. We have labeled the entire dataset automatically using LDA and NMF methods. Now, we are training models to learn the function representation that NMF has provided. For this to work, the underlying assumption is that NMF provided the absolute perfect topic distribution and classification. However, that is dangerous to assume without doing a random sanity check of the dataset. 

df_path = '/content/drive/MyDrive/Wysa/final.pkl'

import pandas as pd

df = pd.read_pickle(df_path)

df.head()

First, we create our train, dev, and test sets with the labels as the NMF topics.

from sklearn.model_selection import train_test_split

xtrain, xtest, ytrain, ytest = train_test_split(df['lemma_text'], df['nmf_topics'], test_size=0.1, stratify = df['nmf_topics'])

xtrain, xdev, ytrain, ydev = train_test_split(xtrain, ytrain, test_size=0.1, stratify=ytrain)

print(xtrain.shape)
print(xdev.shape)
print(xtest.shape)

We create a basic pipeline for Multinomial Naive Bayes followed by a grid search over three parameters.

from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.pipeline import Pipeline
from sklearn.metrics import classification_report
from sklearn.model_selection import GridSearchCV

naive_bayes = Pipeline([('vect', CountVectorizer(stop_words='english')),
               ('tfidf', TfidfTransformer()),
               ('clf', MultinomialNB(fit_prior=False)),
              ], verbose=True)

parameters = {'vect__ngram_range': [(1, 2), (1, 3)],
               'tfidf__use_idf': [True],
               'clf__alpha': (1, 1e-2)}

gs_naive_bayes = GridSearchCV(naive_bayes, parameters, verbose=3)

gs_naive_bayes = gs_naive_bayes.fit(xtrain, ytrain)

gs_naive_bayes.best_params_

ydev_pred = gs_naive_bayes.predict(xdev)
print(classification_report(ydev, ydev_pred))

ytest_pred = gs_naive_bayes.predict(xtest)
print(classification_report(ytest, ytest_pred))

63% accuracy is a decent baseline. Since there is no severe class imbalance either, the FScores are also quite decent, barring class 3.

We do the same for LDA topics.

xtrainlda, xtestlda, ytrainlda, ytestlda = train_test_split(df['lemma_title'], df['lda_topics'], test_size=0.1, stratify = df['lda_topics'])
xtrainlda, xdevlda, ytrainlda, ydevlda = train_test_split(xtrainlda, ytrainlda, test_size=0.1, stratify=ytrainlda)

naive_bayes_lda = Pipeline([('vect', CountVectorizer(stop_words='english')),
               ('tfidf', TfidfTransformer()),
               ('clf', MultinomialNB(fit_prior=False)),
              ], verbose=True)

parameters_lda = {'vect__ngram_range': [(1, 2), (1, 3)],
               'tfidf__use_idf': [True],
               'clf__alpha': (1, 1e-2)}

gs_naive_bayes_lda = GridSearchCV(naive_bayes_lda, parameters_lda, verbose=3)

gs_naive_bayes_lda = gs_naive_bayes_lda.fit(xtrainlda, ytrainlda)

gs_naive_bayes_lda.best_params_

ydev_pred_lda = gs_naive_bayes_lda.predict(xdevlda)
print(classification_report(ydevlda, ydev_pred_lda))

ytest_pred_lda = gs_naive_bayes_lda.predict(xtestlda)
print(classification_report(ytestlda, ytest_pred_lda))

LDA topic classification seems to be a bit of a bust.


Now that we have established a classical ML baseline, lets also get a neural baseline model for classifying the NMF topics.

!pip install transformers

import torch
from tqdm import tqdm
from transformers import DistilBertTokenizerFast

tokenizer = DistilBertTokenizerFast.from_pretrained('distilbert-base-uncased')
train_encodings = tokenizer(list(xtrain), truncation=True, padding=True)
val_encodings = tokenizer(list(xdev), truncation=True, padding=True)

class IndiaDataset(torch.utils.data.Dataset):
    def __init__(self, encodings, labels):
        self.encodings = encodings
        self.labels = labels

    def __getitem__(self, idx):
        item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
        item['labels'] = torch.tensor(self.labels[idx])
        return item

    def __len__(self):
        return len(self.labels)

train_dataset = IndiaDataset(train_encodings, list(ytrain))
val_dataset = IndiaDataset(val_encodings, list(ydev))

from transformers import DistilBertForSequenceClassification, BertConfig

config = BertConfig.from_pretrained('distilbert-base-uncased')
config.num_labels = 5
model = DistilBertForSequenceClassification(config)
model.parameters

from torch.utils.data import DataLoader
from transformers import DistilBertForSequenceClassification, AdamW

device = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
# device = torch.device('cpu')
print(device)

# model = DistilBertForSequenceClassification.from_pretrained('distilbert-base-uncased')
model.to(device)


train_loader = DataLoader(train_dataset, batch_size=16, shuffle=True)
val_loader = DataLoader(val_dataset, batch_size=16, shuffle=True)
model.train()
optim = AdamW(model.parameters(), lr=5e-5)

for epoch in range(5):
    all_train_preds = None
    all_train_labels = None
    all_val_preds = None
    all_val_labels = None
    model.train()
    for batch in tqdm(train_loader):
        optim.zero_grad()
        input_ids = batch['input_ids'].to(device)
        attention_mask = batch['attention_mask'].to(device)
        labels = batch['labels'].to(device)
        outputs = model(input_ids, attention_mask=attention_mask, labels=labels)
        loss = outputs[0]
        logits = outputs[1]
        preds = torch.argmax(logits,dim=1)

        if all_train_preds is None:
            all_train_preds = preds
            all_train_labels = labels
        else:
            all_train_preds = torch.cat((all_train_preds,preds))
            all_train_labels = torch.cat((all_train_labels,labels))

        loss.backward()
        optim.step()
    
    print("train accuracy:", torch.sum(all_train_preds==all_train_labels)/all_train_labels.shape[0])
    
model.eval()
for batch in tqdm(val_loader):
    input_ids = batch['input_ids'].to(device)
    attention_mask = batch['attention_mask'].to(device)
    labels = batch['labels'].to(device)
    outputs = model(input_ids, attention_mask=attention_mask, labels=labels)
    val_loss = outputs[0]
    logits = outputs[1]
    preds = torch.argmax(logits,dim=1)

    if all_val_preds is None:
        all_val_preds = preds
        all_val_labels = labels
    else:
        all_val_preds = torch.cat((all_val_preds,preds))
        all_val_labels = torch.cat((all_val_labels,labels))

print("val accuracy:", torch.sum(all_val_preds==all_val_labels)/all_val_labels.shape[0])

print(classification_report(all_val_labels.cpu(), all_val_preds.cpu()))

Indeed, we improve quite a bit on our Naive Bayes outcome. Better preprocessing, trying out different pre-trained models more suited to reddit, and hyperparameter tuning should be able to increase our FScores further. However, before doing that, we need to establish the validity of our labels in a better way. In the interest of time, I am concluding this task at this stage. Thank you :)

torch.save(model.state_dict(), '/content/drive/MyDrive/Wysa/distilbert_nmf_topics.pt')
